K-D Tree Implementation Request (#256)
* File Creation

Initial creation of the implementation and test files.

* Tree Creation

The basic structure of the k-dimension tree has been implemented, with stand-in methods for insert, search, findMin, and delete.

* Insert Method Added

The insert method, as far as can be told without the search method, is functioning properly.

* Search Method Added

Implemented better tests for insert method and completed the search method, with tests.

* Updated Tests

Reformatted tests, and added a new test to account for bad insert input.

* Insert Method Updated

Fixed a coding error where null coordinate sets are allowed into the k-dimensional tree. A test for this case has also been added.

* Robust Changes

Added dimension checks for tree creation and for the search method.

* Dimensions Working

The previous update treated every tree as one with zero or less dimensions. This update fixes that.

* Added KDNode Comparison

This function should make it easier to compare two nodes on a given axis. This is to set groundwork for the FindMin method.

* Generic Implementation

Methods changed from int implementation to generic implementation, limited to comparables.

* FindMin and Comments Added

The findMin method was added successfully. Comments were also added to make the code more readable.

* Delete Method Added

Added delete method with tests. The delete method is capable of handling root deletes as well as general and bad input cases.

* Test Names Organized

The names of the tests have been edited to make them easier to find on the index.html page.

* Tests Finalized

Updated the exception tests and commented longer tests for better readability. Also deleted unused imports.

* Final Formatted Product

Recursive methods have been made private and the code has been reformatted to comply with the Google Style Guide. All tests run successfully to completion. The implementation is complete.