Use relative links inside repository (#279)
**WHAT:**
Use `relative` links in `Readme.md` instead of `absolute` links.

**WHY:**
When you `fork` the repository it still routes to the original repo from the `Readme` file because of the `absolute` paths. 
Another thing is that it always points to `master` branch regardless of which branch you are in.

**HOW:**
Using `relative` URL paths will fix it and always route you to the same repository/branch in which you are in.