Fixed articulation points alg at the root node
In the current implementation, the noot node is always marked as an articulation point but this is not correct. To determine if the root node is an articulation point, an additional variable called "components" is introduced. This variable counts the number of connected components connected to the root node in the case where the root node is removed. If it is more than 1, that means that the root node would split the graph.